Move DatastoreIO to io.datastore package
